fix(formulas) Create formula names that are valid identifiers
